--- src/Components/Form/Container/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./container;
