+++ src/Sections/AddPatient/add-patient.tsx
@@ -1,12 +1,11 @@
 import TextInput from "../../Components/Form/InputText/text-input";
 import Label from "../../Components/Form/Label/label";
-import Container from ".../../Components/Form/Container/container;
-import NewPacient from ".../../Components/Form/FormTemplates/NewPacient;
+import Container from "../../Components/Form/Container/ContainerForminerForm";
+import NewPacient from "../../Components/Form/FormTemplates/NewPacient";
 
 import { Stack } from "@chakra-ui/react";
 
 const AddPatient = () => {
-  
   const nombre = {
     name: "inputNombre",
     placeholder: "Nombre completo",
@@ -18,11 +17,9 @@ const AddPatient = () => {
     label: <Label value="6/09/1996" />,
   };
 
-  
-
   return (
     <Stack w="100%" alignItems="center" justifyContent="center">
-      <Container childComp={NewPacient} />
+      <ContainerForm> <NewPacient/> </ContainerForm>
     </Stack>
   );
 };
